+++ src/app/app.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
-import { AuthModule } from 'src/auth/auth.module';
+import { AuthModule } from 'src/auth';
 import { PostModule } from 'src/post';
-import { UserModule } from 'src/user/user.module';
+import { S3Module } from 'src/s3';
+import { UserModule } from 'src/user';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { S3Module } from 'src/s3/s3.module';
 
 @Module({
   imports: [UserModule, PostModule, AuthModule, S3Module],
