+++ src/post/post.module.ts
@@ -1,13 +1,12 @@
 import { Module } from '@nestjs/common';
 import { PrismaService } from 'src/prisma';
 import { UserService } from 'src/user';
-import { PostController } from './post.controller';
-import { PostService } from './post.service';
-import { S3Module } from 'src/s3/s3.module';
+import { PostService, PostController } from 'src/post';
+import { S3Module } from 'src/s3';
 
 @Module({
   controllers: [PostController],
   providers: [PostService, PrismaService, UserService],
-  imports: [S3Module]
+  imports: [S3Module],
 })
 export class PostModule {}
