+++ src/s3/index.ts
@@ -1,0 +1,2 @@
+export * from './s3.module';
+export * from './s3.service';
